# Defer eventFd / timerFd reads until end of loop.