Motivation:

When submitting a task an epoll event loop, the submitter notifies the
loop of the new task by writing to an event fd.  If the loop is waiting
in epoll_wait, it wakes up, reads the eventFd, and processes the tasks.
Similarly when a timer fires, the loop wakes up, reads the timerFd, and
processes the tasks.

This is a slight latency hit, because the submitter (and timer) want
their tasks to be processed first, rather than reading from the fds.
The reads cost a JNI hop and a syscall, which my unscientific
measurements put at around 500ns-1000ns.

Modification:

Change the epoll loop to defer reading of the eventFd and timerFd.  The
fds are read after all tasks have been processed.

Result:

More timely handling of tasks.

Notes:

This only really helps if the event loop has less than 1 active network
tasks.  The benefit is lost as the event loop is more active, because it
naturally is awake more avoiding the need for the eventFd read at wake.
The behavior was discovered in a gRPC Benchmark, where there is only
ever one active RPC at a time, which is started from off the event loop.

Some benchmark results from my slightly noisy Skylake:

```
Before:

Iteration   1: 155826.732 ops/s
Iteration   2: 156789.940 ops/s
Iteration   3: 155141.115 ops/s
...

After:

Iteration   1: 159719.071 ops/s
Iteration   2: 166250.139 ops/s
Iteration   3: 165861.565 ops/s
```