@normanmaurer 2019-05-26
@netty-bot test this please