@normanmaurer 2019-05-28
@carl-mastrangelo I think we can just close this and apply https://github.com/netty/netty/pull/9192 ... I added you as co-author FTW :)